requirements.txt
altair==5.1.2
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
msgpack-python==0.5.6
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.0.0
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.11.3
service-identity==23.1.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
zope.interface==6.0
